pyproject.toml
[tool.poetry]
name = "itrackcryptobot"
version = "0.1.0"
description = ""
authors = ["Oliver Marshall <olivershawmarshall@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^13.11"
pycoingecko = "^2.2.0"
matplotlib = "^3.5.1"
pytz = "^2022.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
